Editor config with some analyzers.
