+++ ConsoleApp22/INode.cs
@@ -2,6 +2,6 @@
 {
     public interface INode
     {
-        string? GetAttribute(string name, string fallback = null);
+        string? GetAttribute(string name, string? fallback = null);
     }
 }
