+++ Mjml.Net/Components/Body/ColumnComponent.cs
@@ -1,28 +1,15 @@
-﻿namespace Mjml.Net.Components.Body
+﻿using Mjml.Net.Helpers;
+
+namespace Mjml.Net.Components.Body
 {
-    public sealed class SectionComponent : BodyComponentBase
+    public sealed class ColumnComponent : BodyComponentBase
     {
-        public override string ComponentName => "mj-section";
+        public override string ComponentName => "mj-column";
 
         public override void Render(IHtmlRenderer renderer, INode node)
         {
-            RenderChildren(renderer, node);
-        }
-
-        private static void RenderChildren(IHtmlRenderer renderer, INode node)
-        {
-            renderer.RenderChildren(new ChildOptions
-            {
-                Renderer = child =>
-                {
-                    renderer.ElementStart("div")
-                        .Attr("width", node.GetAttribute("width"));
-
-                    child.Render();
-
-                    renderer.ElementEnd("link");
-                }
-            });
+            // Set the breakpoint if not set before.
+            renderer.SetGlobalData("default", Breakpoint.Default, true);
         }
     }
 }
