+++ Mjml.Net/Helpers/Font.cs
@@ -1,6 +1,8 @@
 ﻿namespace Mjml.Net.Helpers
 {
+#pragma warning disable SA1313 // Parameter names should begin with lower-case letter
     public sealed record Font(string Href);
+#pragma warning restore SA1313 // Parameter names should begin with lower-case letter
 
     public sealed class FontHelper : IHelper
     {
