+++ Mjml.Net/Helpers/Preview.cs
@@ -17,7 +17,7 @@
             {
                 renderer.ElementStart("div")
                     .Attr("style", "display:none;font-size:1px;color:#ffffff;line-height:1px;max-height:0px;max-width:0px;opacity:0;overflow:hidden;");
-                
+
                 foreach (var preview in data.Values.OfType<Preview>())
                 {
                     renderer.Content(preview.Value);
